alsaconf:  use 'type -p', not which
Here is a small patch for alsaconf, which removes an useless dependency
on `which' by using a bash built-in instead : `type -p'.  I encountered
the problem of the missing `which' while using alsaconf on a clfs-built
linux system.  It is useless to install `which' when we already have
`type -p', and alsaconf already needs bash, so this does not replace
a dependency by another one, but really removes a dependency for alsaconf.

Signed-off-by: Takashi Iwai <tiwai@suse.de>